authenticate create post and delete post

+++ graphql/resolvers/questions.js
@@ -1,4 +1,7 @@
+const { AuthenticationError, UserInputError } = require('apollo-server')
+
 const Question = require('../../models/Question')
+const auth = require('../../util/auth')
 
 module.exports = {
   Query: {
@@ -24,20 +27,33 @@ module.exports = {
     }
   },
   Mutation: {
-    async createQuestion(_, { body }) {
+    async createQuestion(_, { body }, context) {
+      const user = auth(context)
+
+      if (body.trim() === '') {
+        throw new Error('Question body must not be empty')
+      }
+
       const newQuestion = new Question({
         body,
-        username: 'dimas',
+        email: user.email,
+        user: user.id,
         createdAt: new Date().toISOString()
       })
       const question = await newQuestion.save()
       return question
     },
-    async deleteQuestion(_, { questionId }) {
+    async deleteQuestion(_, { questionId }, context) {
+      const user = auth(context)
+
       try {
         const question = await Question.findById(questionId)
-        await question.delete()
-        return 'Question deleted successfully'
+        if (user.email === question.email) {
+          await question.delete()
+          return 'Question deleted successfully'
+        } else {
+          throw new AuthenticationError('Action not allowed')
+        }
       } catch (err) {
         throw new Error(err)
       }
